# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AIChatMessage:
    properties:
      content:
        type: string
      role:
        enum:
        - user
        - assistant
        - system
        type: string
    required:
    - content
    - role
    type: object
  dto.AIChatRequest:
    properties:
      message:
        type: string
      messages:
        description: 历史消息上下文
        items:
          $ref: '#/definitions/dto.AIChatMessage'
        type: array
      stream:
        type: boolean
    required:
    - message
    type: object
  dto.AIChatResponse:
    properties:
      message:
        type: string
      timestamp:
        type: string
      toolCalls:
        items:
          type: string
        type: array
    type: object
  dto.AIChatStreamResponse:
    properties:
      content:
        type: string
      done:
        type: boolean
    type: object
...
  model.ErrResponseDontShowError:
    description: 错误的API响应标准格式,不展示 error
    properties:
      code:
        example: 400
        type: integer
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  model.SuccessResponse:
    description: 成功的API响应标准格式
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
host: localhost:2333
info:
  contact:
    email: huahua1319873800@outlook.com
    name: API Support
    url: https://github.com/mingrenya/AI-Waf
  description: MRYa 应用防火墙管理系统 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/mingrenya/AI-Waf
  title: MRYa-WAF API
  version: "1.0"
paths:
  /ai/chat:
    post:
      consumes:
      - application/json
      description: 与AI助手进行对话
      parameters:
      - description: 聊天请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AIChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AIChatResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 发送AI聊天消息
      tags:
      - AI
  /ai/chat/stream:
    post:
      consumes:
      - application/json
      description: 流式返回AI聊天响应
      parameters:
      - description: 聊天请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AIChatRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AIChatStreamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 流式AI聊天
      tags:
      - AI
...
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer {token} 格式进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"